Module ic_replica::setup

Functions

create_consensus_pool_dir
Create the consensus pool directory (if none exists)

get_catch_up_package
Parse the catch-up package given via command-line args (if one was given)

get_config_source
Return the source from which to derive the Replica's config

get_replica_logger
Return a ReplicaLogger and its AsyncGuard

get_subnet_id
Return the subnet ID of the given node

get_subnet_type
Return the subnet type of the given subnet.

parse_args
Parse command-line args into ReplicaArgs

set_replica_version
Set the Replica version passed in via command-line

setup_crypto_provider
Setup a crypto provider provided a CryptoConfig, a RegistryClient, a ReplicaLogger, and a MetricsRegistry.

setup_crypto_provider_with_node_id
Similarly to setup_crypto_provider sets up a cryptographic functionality provider. The difference is we indicate a particular node id (NodeId).

setup_crypto_registry